Split AVID impl into batch_avss_avid; restore original batch_avss
Move the new AVID-based implementation into a new batch_avss_avid module
(plus its benchmark) and restore the original batch_avss module verbatim,
along with the complaint module it depends on.

- presigning stays wired to the original batch_avss output types
- e2e tests and benchmarks exercise the new batch_avss_avid implementation,
  converting its (structurally identical) outputs via From impls so they can
  feed presigning
